Show Pod Errors and other pump highlights
